Remove HAVE_LOCALE_H symbol
The locale.h header is part of the C89 standard and is present on
all today's systems already. The HAVE_LOCALE_H symbol is defined
by PHP's build system and relying on it is neither a good practice
neither needed anymore since the locale.h check would always define
it.

http://port70.net/~nsz/c/c89/c89-draft.html#4.1.2